完成首页界面及点击事件的预留接口

+++ EnterpriseEdition/EnterpriseEdition/Class/Util/Util.h
@@ -40,5 +40,5 @@
 /**
  随机生成匹配度
  */
-+(CGFloat)randomRate;
++(NSInteger)randomRate;
 @end
